deletors: skip projected resources
Skip attempting to delete projected resources in the APIBinding and
Workspace Resource Deletors. These projected resources will
automatically go away when their underlying resources are deleted.

(cherry picked from commit 6d3c78a2234b922d75a2c5e121e6b9d8037d0761)

+++ pkg/reconciler/apis/apibindingdeletion/apibinding_deletor.go
@@ -29,6 +29,7 @@ import (
 	"k8s.io/klog/v2"
 
 	apisv1alpha1 "github.com/kcp-dev/kcp/pkg/apis/apis/v1alpha1"
+	"github.com/kcp-dev/kcp/pkg/projection"
 )
 
 type gvrDeletionMetadata struct {
@@ -56,6 +57,13 @@ func (c *Controller) deleteAllCRs(ctx context.Context, apibinding *apisv1alpha1.
 				Group:    resource.Group,
 				Resource: resource.Resource,
 				Version:  version,
+			}
+
+			// Don't try to delete projected resources such as tenancy.kcp.dev v1beta1 Workspaces - these are virtual
+			// projections and we shouldn't try to delete them. The projections will disappear when the real underlying
+			// data (e.g. ClusterWorkspaces) are deleted.
+			if projection.Includes(gvr) {
+				continue
 			}
 
 			deletionMetadata, err := c.deleteAllCR(ctx, logicalcluster.From(apibinding), gvr)
